+++ noodles-sam/src/header/parser.rs
@@ -20,8 +20,6 @@ use super::{
 pub enum ParseError {
     /// A header record is not on the first line.
     UnexpectedHeader,
-    /// The header record is invalid.
-    InvalidHeaderRecord(super::record::ParseError),
     /// The record is invalid.
     InvalidRecord(record::ParseError),
     /// A reference sequence name is duplicated.
@@ -37,7 +35,6 @@ pub enum ParseError {
 impl error::Error for ParseError {
     fn source(&self) -> Option<&(dyn error::Error + 'static)> {
         match self {
-            Self::InvalidHeaderRecord(e) => Some(e),
             Self::InvalidRecord(e) => Some(e),
             _ => None,
         }
@@ -48,7 +45,6 @@ impl fmt::Display for ParseError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
             Self::UnexpectedHeader => f.write_str("unexpected @HD"),
-            Self::InvalidHeaderRecord(_) => f.write_str("invalid header record"),
             Self::InvalidRecord(_) => f.write_str("invalid record"),
             Self::DuplicateReferenceSequenceName(name) => {
                 write!(f, "duplicate reference sequence name: {name}")