+++ noodles-sam/src/header/record/value/map/header.rs
@@ -12,11 +12,10 @@ pub use self::{
     version::Version,
 };
 
-use std::{error, fmt};
+use std::fmt;
 
 use self::{builder::Builder, tag::StandardTag};
-use super::{Fields, Inner, Map, OtherFields};
-use crate::header::parser::Context;
+use super::{Inner, Map, OtherFields};
 
 /// A SAM header record header map value.
 ///
@@ -194,143 +193,6 @@ impl fmt::Display for Map<Header> {
     }
 }
 
-/// An error returned when a raw header header record fails to parse.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum ParseError {
-    /// A field is missing.
-    MissingField(Tag),
-    /// A tag is invalid.
-    InvalidTag(super::tag::ParseError),
-    /// A tag is duplicated.
-    DuplicateTag(Tag),
-    /// The version is invalid.
-    InvalidVersion(version::ParseError),
-    /// The sort order is invalid.
-    InvalidSortOrder(sort_order::ParseError),
-    /// The group order is invalid.
-    InvalidGroupOrder(group_order::ParseError),
-    /// The subsort order is invalid.
-    InvalidSubsortOrder(subsort_order::ParseError),
-}
-
-impl error::Error for ParseError {
-    fn source(&self) -> Option<&(dyn error::Error + 'static)> {
-        match self {
-            Self::InvalidTag(e) => Some(e),
-            Self::InvalidVersion(e) => Some(e),
-            Self::InvalidSortOrder(e) => Some(e),
-            Self::InvalidGroupOrder(e) => Some(e),
-            Self::InvalidSubsortOrder(e) => Some(e),
-            _ => None,
-        }
-    }
-}
-
-impl fmt::Display for ParseError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::MissingField(tag) => write!(f, "missing field: {tag}"),
-            Self::InvalidTag(_) => write!(f, "invalid tag"),
-            Self::DuplicateTag(tag) => write!(f, "duplicate tag: {tag}"),
-            Self::InvalidVersion(_) => write!(f, "invalid version"),
-            Self::InvalidSortOrder(_) => write!(f, "invalid sort order"),
-            Self::InvalidGroupOrder(_) => write!(f, "invalid group order"),
-            Self::InvalidSubsortOrder(_) => write!(f, "invalid subsort order"),
-        }
-    }
-}
-
-impl TryFrom<Fields> for Map<Header> {
-    type Error = ParseError;
-
-    fn try_from(fields: Fields) -> Result<Self, Self::Error> {
-        Self::try_from((&Context::default(), fields))
-    }
-}
-
-impl TryFrom<(&Context, Fields)> for Map<Header> {
-    type Error = ParseError;
-
-    fn try_from((ctx, fields): (&Context, Fields)) -> Result<Self, Self::Error> {
-        let mut version = None;
-        let mut sort_order = None;
-        let mut group_order = None;
-        let mut subsort_order = None;
-
-        let mut other_fields = OtherFields::new();
-
-        for (key, value) in fields {
-            let tag = key.parse().map_err(ParseError::InvalidTag)?;
-
-            match tag {
-                tag::VERSION => parse_version(&value)
-                    .and_then(|v| try_replace(&mut version, ctx, tag::VERSION, v))?,
-                tag::SORT_ORDER => parse_sort_order(&value)
-                    .and_then(|v| try_replace(&mut sort_order, ctx, tag::SORT_ORDER, v))?,
-                tag::GROUP_ORDER => parse_group_order(&value)
-                    .and_then(|v| try_replace(&mut group_order, ctx, tag::GROUP_ORDER, v))?,
-                tag::SUBSORT_ORDER => parse_subsort_order(&value)
-                    .and_then(|v| try_replace(&mut subsort_order, ctx, tag::SUBSORT_ORDER, v))?,
-                Tag::Other(t) => try_insert(&mut other_fields, ctx, t, value)?,
-            }
-        }
-
-        let version = version.ok_or(ParseError::MissingField(tag::VERSION))?;
-
-        Ok(Self {
-            inner: Header {
-                version,
-                sort_order,
-                group_order,
-                subsort_order,
-            },
-            other_fields,
-        })
-    }
-}
-
-fn parse_version(s: &str) -> Result<Version, ParseError> {
-    s.parse().map_err(ParseError::InvalidVersion)
-}
-
-fn parse_sort_order(s: &str) -> Result<SortOrder, ParseError> {
-    s.parse().map_err(ParseError::InvalidSortOrder)
-}
-
-fn parse_group_order(s: &str) -> Result<GroupOrder, ParseError> {
-    s.parse().map_err(ParseError::InvalidGroupOrder)
-}
-
-fn parse_subsort_order(s: &str) -> Result<SubsortOrder, ParseError> {
-    s.parse().map_err(ParseError::InvalidSubsortOrder)
-}
-
-fn try_replace<T>(
-    option: &mut Option<T>,
-    ctx: &Context,
-    tag: Tag,
-    value: T,
-) -> Result<(), ParseError> {
-    if option.replace(value).is_some() && !ctx.allow_duplicate_tags() {
-        Err(ParseError::DuplicateTag(tag))
-    } else {
-        Ok(())
-    }
-}
-
-fn try_insert(
-    other_fields: &mut OtherFields<StandardTag>,
-    ctx: &Context,
-    tag: super::tag::Other<StandardTag>,
-    value: String,
-) -> Result<(), ParseError> {
-    if other_fields.insert(tag, value).is_some() && !ctx.allow_duplicate_tags() {
-        Err(ParseError::DuplicateTag(Tag::Other(tag)))
-    } else {
-        Ok(())
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -357,13 +219,5 @@ mod tests {
         assert_eq!(header.to_string(), "VN:1.6\tSO:unsorted\tGO:query");
 
         Ok(())
-    }
-
-    #[test]
-    fn test_try_from_fields_for_map_header_with_missing_version() {
-        assert_eq!(
-            Map::<Header>::try_from(vec![(String::from("SO"), String::from("coordinate"))]),
-            Err(ParseError::MissingField(tag::VERSION))
-        );
     }
 }