+++ noodles-sam/src/header/record/value/map/read_group.rs
@@ -7,11 +7,10 @@ pub(crate) mod tag;
 pub use self::platform::Platform;
 pub(crate) use self::tag::Tag;
 
-use std::{error, fmt, num};
+use std::fmt;
 
 use self::{builder::Builder, tag::StandardTag};
-use super::{Fields, Inner, Map, OtherFields};
-use crate::header::parser::Context;
+use super::{Inner, Map};
 
 /// A SAM header record read group map value.
 ///
@@ -274,167 +273,6 @@ impl fmt::Display for Map<ReadGroup> {
     }
 }
 
-/// An error returned when a raw header read group record fails to parse.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum ParseError {
-    /// A field is missing.
-    MissingField(Tag),
-    /// A tag is invalid.
-    InvalidTag(super::tag::ParseError),
-    /// A tag is duplicated.
-    DuplicateTag(Tag),
-    /// The predicted median insert size is invalid.
-    InvalidPredictedMedianInsertSize(num::ParseIntError),
-    /// The platform is invalid.
-    InvalidPlatform(platform::ParseError),
-}
-
-impl error::Error for ParseError {
-    fn source(&self) -> Option<&(dyn error::Error + 'static)> {
-        match self {
-            Self::InvalidTag(e) => Some(e),
-            Self::InvalidPredictedMedianInsertSize(e) => Some(e),
-            Self::InvalidPlatform(e) => Some(e),
-            _ => None,
-        }
-    }
-}
-
-impl fmt::Display for ParseError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::MissingField(tag) => write!(f, "missing field: {tag}"),
-            Self::InvalidTag(_) => write!(f, "invalid tag"),
-            Self::DuplicateTag(tag) => write!(f, "duplicate tag: {tag}"),
-            Self::InvalidPredictedMedianInsertSize(_) => {
-                write!(f, "invalid predicted median insert size")
-            }
-            Self::InvalidPlatform(_) => write!(f, "invalid platform"),
-        }
-    }
-}
-
-impl TryFrom<Fields> for Map<ReadGroup> {
-    type Error = ParseError;
-
-    fn try_from(fields: Fields) -> Result<Self, Self::Error> {
-        Self::try_from((&Context::default(), fields))
-    }
-}
-
-impl TryFrom<(&Context, Fields)> for Map<ReadGroup> {
-    type Error = ParseError;
-
-    fn try_from((ctx, fields): (&Context, Fields)) -> Result<Self, Self::Error> {
-        let mut barcode = None;
-        let mut sequencing_center = None;
-        let mut description = None;
-        let mut produced_at = None;
-        let mut flow_order = None;
-        let mut key_sequence = None;
-        let mut library = None;
-        let mut program = None;
-        let mut predicted_median_insert_size = None;
-        let mut platform = None;
-        let mut platform_model = None;
-        let mut platform_unit = None;
-        let mut sample = None;
-
-        let mut other_fields = OtherFields::new();
-
-        for (key, value) in fields {
-            let tag = key.parse().map_err(ParseError::InvalidTag)?;
-
-            match tag {
-                tag::ID => return Err(ParseError::DuplicateTag(tag::ID)),
-                tag::BARCODE => try_replace(&mut barcode, ctx, tag::BARCODE, value)?,
-                tag::SEQUENCING_CENTER => {
-                    try_replace(&mut sequencing_center, ctx, tag::SEQUENCING_CENTER, value)?
-                }
-                tag::DESCRIPTION => try_replace(&mut description, ctx, tag::DESCRIPTION, value)?,
-                tag::PRODUCED_AT => try_replace(&mut produced_at, ctx, tag::PRODUCED_AT, value)?,
-                tag::FLOW_ORDER => try_replace(&mut flow_order, ctx, tag::FLOW_ORDER, value)?,
-                tag::KEY_SEQUENCE => try_replace(&mut key_sequence, ctx, tag::KEY_SEQUENCE, value)?,
-                tag::LIBRARY => try_replace(&mut library, ctx, tag::LIBRARY, value)?,
-                tag::PROGRAM => try_replace(&mut program, ctx, tag::PROGRAM, value)?,
-                tag::PREDICTED_MEDIAN_INSERT_SIZE => parse_predicted_median_insert_size(&value)
-                    .and_then(|v| {
-                        try_replace(
-                            &mut predicted_median_insert_size,
-                            ctx,
-                            tag::PREDICTED_MEDIAN_INSERT_SIZE,
-                            v,
-                        )
-                    })?,
-                tag::PLATFORM => parse_platform(&value)
-                    .and_then(|v| try_replace(&mut platform, ctx, tag::PLATFORM, v))?,
-                tag::PLATFORM_MODEL => {
-                    try_replace(&mut platform_model, ctx, tag::PLATFORM_MODEL, value)?
-                }
-                tag::PLATFORM_UNIT => {
-                    try_replace(&mut platform_unit, ctx, tag::PLATFORM_UNIT, value)?
-                }
-                tag::SAMPLE => try_replace(&mut sample, ctx, tag::SAMPLE, value)?,
-                Tag::Other(t) => try_insert(&mut other_fields, ctx, t, value)?,
-            }
-        }
-
-        Ok(Self {
-            inner: ReadGroup {
-                barcode,
-                sequencing_center,
-                description,
-                produced_at,
-                flow_order,
-                key_sequence,
-                library,
-                program,
-                predicted_median_insert_size,
-                platform,
-                platform_model,
-                platform_unit,
-                sample,
-            },
-            other_fields,
-        })
-    }
-}
-
-fn parse_predicted_median_insert_size(s: &str) -> Result<i32, ParseError> {
-    s.parse()
-        .map_err(ParseError::InvalidPredictedMedianInsertSize)
-}
-
-fn parse_platform(s: &str) -> Result<Platform, ParseError> {
-    s.parse().map_err(ParseError::InvalidPlatform)
-}
-
-fn try_replace<T>(
-    option: &mut Option<T>,
-    ctx: &Context,
-    tag: Tag,
-    value: T,
-) -> Result<(), ParseError> {
-    if option.replace(value).is_some() && !ctx.allow_duplicate_tags() {
-        Err(ParseError::DuplicateTag(tag))
-    } else {
-        Ok(())
-    }
-}
-
-fn try_insert(
-    other_fields: &mut OtherFields<StandardTag>,
-    ctx: &Context,
-    tag: super::tag::Other<StandardTag>,
-    value: String,
-) -> Result<(), ParseError> {
-    if other_fields.insert(tag, value).is_some() && !ctx.allow_duplicate_tags() {
-        Err(ParseError::DuplicateTag(Tag::Other(tag)))
-    } else {
-        Ok(())
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -450,41 +288,5 @@ mod tests {
         assert_eq!(read_group.to_string(), "\tPG:noodles\tPL:ILLUMINA");
 
         Ok(())
-    }
-
-    #[test]
-    fn test_try_from_fields_for_map_read_group_with_duplicate_id() {
-        let fields = vec![(String::from("ID"), String::from("rg0"))];
-
-        assert_eq!(
-            Map::<ReadGroup>::try_from(fields),
-            Err(ParseError::DuplicateTag(tag::ID))
-        );
-    }
-
-    #[test]
-    fn test_try_from_fields_for_map_read_group_with_an_invalid_predicted_median_insert_size() {
-        let fields = vec![
-            (String::from("PG"), String::from("noodles")),
-            (String::from("PI"), String::from("unknown")),
-        ];
-
-        assert!(matches!(
-            Map::<ReadGroup>::try_from(fields),
-            Err(ParseError::InvalidPredictedMedianInsertSize(_))
-        ));
-    }
-
-    #[test]
-    fn test_try_from_fields_for_map_read_group_with_an_invalid_platform() {
-        let fields = vec![
-            (String::from("PG"), String::from("noodles")),
-            (String::from("PL"), String::from("unknown")),
-        ];
-
-        assert!(matches!(
-            Map::<ReadGroup>::try_from(fields),
-            Err(ParseError::InvalidPlatform(_))
-        ));
     }
 }