+++ noodles-sam/src/header/record/value/map/reference_sequence.rs
@@ -8,10 +8,7 @@ pub mod molecule_topology;
 pub mod name;
 pub(crate) mod tag;
 
-use std::{
-    error, fmt,
-    num::{self, NonZeroUsize},
-};
+use std::{fmt, num::NonZeroUsize};
 
 pub(crate) use self::tag::Tag;
 pub use self::{
@@ -20,8 +17,7 @@ pub use self::{
 };
 
 use self::{builder::Builder, tag::StandardTag};
-use super::{Fields, Inner, Map, OtherFields};
-use crate::header::parser::Context;
+use super::{Inner, Map, OtherFields};
 
 /// A SAM header record reference sequence map value.
 ///
@@ -299,175 +295,6 @@ impl fmt::Display for Map<ReferenceSequence> {
     }
 }
 
-/// An error returned when a raw header reference sequence record fails to parse.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum ParseError {
-    /// A field is missing.
-    MissingField(Tag),
-    /// A tag is invalid.
-    InvalidTag(super::tag::ParseError),
-    /// A tag is duplicated.
-    DuplicateTag(Tag),
-    /// The name is invalid.
-    InvalidName(name::ParseError),
-    /// The length is invalid.
-    InvalidLength(num::ParseIntError),
-    /// The alternative locus is invalid.
-    InvalidAlternativeLocus(alternative_locus::ParseError),
-    /// The alternative names is invalid.
-    InvalidAlternativeNames(alternative_names::ParseError),
-    /// The MD5 checksum is invalid.
-    InvalidMd5Checksum(md5_checksum::ParseError),
-    /// The molecule topology is invalid.
-    InvalidMoleculeToplogy(molecule_topology::ParseError),
-}
-
-impl error::Error for ParseError {
-    fn source(&self) -> Option<&(dyn error::Error + 'static)> {
-        match self {
-            Self::InvalidTag(e) => Some(e),
-            Self::InvalidName(e) => Some(e),
-            Self::InvalidLength(e) => Some(e),
-            Self::InvalidAlternativeLocus(e) => Some(e),
-            Self::InvalidAlternativeNames(e) => Some(e),
-            Self::InvalidMd5Checksum(e) => Some(e),
-            Self::InvalidMoleculeToplogy(e) => Some(e),
-            _ => None,
-        }
-    }
-}
-
-impl fmt::Display for ParseError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::MissingField(tag) => write!(f, "missing field: {tag}"),
-            Self::InvalidTag(_) => write!(f, "invalid tag"),
-            Self::DuplicateTag(tag) => write!(f, "duplicate tag: {tag}"),
-            Self::InvalidName(_) => write!(f, "invalid name"),
-            Self::InvalidLength(_) => write!(f, "invalid length"),
-            Self::InvalidAlternativeLocus(_) => write!(f, "invalid alternative locus"),
-            Self::InvalidAlternativeNames(_) => write!(f, "invalid alternative names"),
-            Self::InvalidMd5Checksum(_) => write!(f, "invalid MD5 checksum"),
-            Self::InvalidMoleculeToplogy(_) => write!(f, "invalid molecule topology"),
-        }
-    }
-}
-
-impl TryFrom<Fields> for Map<ReferenceSequence> {
-    type Error = ParseError;
-
-    fn try_from(fields: Fields) -> Result<Self, Self::Error> {
-        Self::try_from((&Context::default(), fields))
-    }
-}
-
-impl TryFrom<(&Context, Fields)> for Map<ReferenceSequence> {
-    type Error = ParseError;
-
-    fn try_from((ctx, fields): (&Context, Fields)) -> Result<Self, Self::Error> {
-        let mut length = None;
-        let mut alternative_locus = None;
-        let mut alternative_names = None;
-        let mut assembly_id = None;
-        let mut description = None;
-        let mut md5_checksum = None;
-        let mut species = None;
-        let mut molecule_topology = None;
-        let mut uri = None;
-
-        let mut other_fields = OtherFields::new();
-
-        for (key, value) in fields {
-            let tag = key.parse().map_err(ParseError::InvalidTag)?;
-
-            match tag {
-                tag::NAME => return Err(ParseError::DuplicateTag(tag::NAME)),
-                tag::LENGTH => parse_length(&value)
-                    .and_then(|v| try_replace(&mut length, ctx, tag::LENGTH, v))?,
-                tag::ALTERNATIVE_LOCUS => parse_alternative_locus(&value).and_then(|v| {
-                    try_replace(&mut alternative_locus, ctx, tag::ALTERNATIVE_LOCUS, v)
-                })?,
-                tag::ALTERNATIVE_NAMES => parse_alternative_names(&value).and_then(|v| {
-                    try_replace(&mut alternative_names, ctx, tag::ALTERNATIVE_NAMES, v)
-                })?,
-                tag::ASSEMBLY_ID => try_replace(&mut assembly_id, ctx, tag::ASSEMBLY_ID, value)?,
-                tag::DESCRIPTION => try_replace(&mut description, ctx, tag::DESCRIPTION, value)?,
-                tag::MD5_CHECKSUM => parse_md5_checksum(&value)
-                    .and_then(|v| try_replace(&mut md5_checksum, ctx, tag::MD5_CHECKSUM, v))?,
-                tag::SPECIES => try_replace(&mut species, ctx, tag::SPECIES, value)?,
-                tag::MOLECULE_TOPOLOGY => parse_molecule_topology(&value).and_then(|v| {
-                    try_replace(&mut molecule_topology, ctx, tag::MOLECULE_TOPOLOGY, v)
-                })?,
-                tag::URI => try_replace(&mut uri, ctx, tag::URI, value)?,
-                Tag::Other(t) => try_insert(&mut other_fields, ctx, t, value)?,
-            }
-        }
-
-        let length = length.ok_or(ParseError::MissingField(tag::LENGTH))?;
-
-        Ok(Self {
-            inner: ReferenceSequence {
-                length,
-                alternative_locus,
-                alternative_names,
-                assembly_id,
-                description,
-                md5_checksum,
-                species,
-                molecule_topology,
-                uri,
-            },
-            other_fields,
-        })
-    }
-}
-
-fn parse_length(s: &str) -> Result<NonZeroUsize, ParseError> {
-    s.parse().map_err(ParseError::InvalidLength)
-}
-
-fn parse_alternative_locus(s: &str) -> Result<AlternativeLocus, ParseError> {
-    s.parse().map_err(ParseError::InvalidAlternativeLocus)
-}
-
-fn parse_alternative_names(s: &str) -> Result<AlternativeNames, ParseError> {
-    s.parse().map_err(ParseError::InvalidAlternativeNames)
-}
-
-fn parse_md5_checksum(s: &str) -> Result<Md5Checksum, ParseError> {
-    s.parse().map_err(ParseError::InvalidMd5Checksum)
-}
-
-fn parse_molecule_topology(s: &str) -> Result<MoleculeTopology, ParseError> {
-    s.parse().map_err(ParseError::InvalidMoleculeToplogy)
-}
-
-fn try_replace<T>(
-    option: &mut Option<T>,
-    ctx: &Context,
-    tag: Tag,
-    value: T,
-) -> Result<(), ParseError> {
-    if option.replace(value).is_some() && !ctx.allow_duplicate_tags() {
-        Err(ParseError::DuplicateTag(tag))
-    } else {
-        Ok(())
-    }
-}
-
-fn try_insert(
-    other_fields: &mut OtherFields<StandardTag>,
-    ctx: &Context,
-    tag: super::tag::Other<StandardTag>,
-    value: String,
-) -> Result<(), ParseError> {
-    if other_fields.insert(tag, value).is_some() && !ctx.allow_duplicate_tags() {
-        Err(ParseError::DuplicateTag(Tag::Other(tag)))
-    } else {
-        Ok(())
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -488,40 +315,5 @@ mod tests {
         );
 
         Ok(())
-    }
-
-    #[test]
-    fn test_try_from_fields_for_map_reference_sequence_with_duplicate_name() {
-        let fields = vec![(String::from("SN"), String::from("sq0"))];
-
-        assert_eq!(
-            Map::<ReferenceSequence>::try_from(fields),
-            Err(ParseError::DuplicateTag(tag::NAME))
-        );
-    }
-
-    #[test]
-    fn test_try_from_fields_for_map_reference_sequence_with_missing_length() {
-        assert_eq!(
-            Map::<ReferenceSequence>::try_from(vec![]),
-            Err(ParseError::MissingField(tag::LENGTH))
-        );
-    }
-
-    #[test]
-    fn test_try_from_fields_for_map_reference_sequence_with_invalid_length() {
-        let fields = vec![(String::from("LN"), String::from("NA"))];
-
-        assert!(matches!(
-            Map::<ReferenceSequence>::try_from(fields),
-            Err(ParseError::InvalidLength(_))
-        ));
-
-        let fields = vec![(String::from("LN"), String::from("0"))];
-
-        assert!(matches!(
-            Map::<ReferenceSequence>::try_from(fields),
-            Err(ParseError::InvalidLength(_))
-        ));
     }
 }